% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\docType{data}
\name{earthquakes_raw}
\alias{earthquakes_raw}
\title{Earthquakes Raw Dataframe}
\format{
An object of class \code{"tibble"}.
}
\usage{
data(earthquakes_raw)
}
\description{
Data provided by the JHU for the Coursera Capstone Project. It contains an exhaustive list of earthquakes registered in our planet.
}
\examples{
data(earthquakes_raw)
}
\keyword{datasets}
