% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(eq_raw)
}
\arguments{
\item{eq_raw}{a raw dataframe from the NOAA Significant Earthquakes data set}
}
\value{
a dataframe with the cleaned location names, numeric lat/long and DATE column
}
\description{
Function that tidies up the raw earthquake data
}
\details{
The function first replaces NA values in month and day with 1 to give a complete date
    it then mutates the column like this
    1. creates a new Date class column called "DATE" by combining the YEAR, MONTH and DAY field
    2. Converts the LATTITUDE and LONGITUDE columns to numeric
    3. Calls eq_location_clean to strip the country name from the location column
}
\examples{
\dontrun{
eq_clean_data(eq_raw)
}

}
