% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{a clean(with eq_clean_data) dataframe from the NOAA Significant Earthquakes data set}
}
\value{
a string containing the html code
}
\description{
creates the HTML to create the popup used in Leaflet
}
\details{
The function creates an html chunk wiht labels and values to represent the Location, Total Deaths and
   Magnitude.  If any of these three is NA, the value and it's corresponding label
   is omitted from the returned string
}
\examples{
\dontrun{
eq_Mexico <- dplyr::filter(eq_clean,COUNTRY == "MEXICO" & lubridate::year(DATE) >2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.))
}

}
