% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{a dataframe with a Location column that contains Country information}
}
\value{
a dataframe with the Country and any folowing colons or semi-colons
    removed from the front of the Location
}
\description{
Function that removes the country name from the location data
}
\details{
The function first removes Country from location appending : and ;
    then trims both ends of the location string
    str_to_title is then used to convert the location to initial upper case
}
\examples{
\dontrun{
eq_location_clean(eq_raw)
}

}
