% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_read_data}
\alias{eq_read_data}
\title{Mastering Software Development in R Specialization Capstone Project
Coursera Capstone Project
Read Earthquakes data.}
\usage{
eq_read_data(filename)
}
\arguments{
\item{filename}{A caracter that contains the name of the file with its respective
extension.}
}
\value{
This function returns the read data in a "dataframe", "tbl_df", "tbl" format.
}
\description{
Mastering Software Development in R Specialization Capstone Project
Coursera Capstone Project
Read Earthquakes data.
}
\note{
This function will generate an error if the filemane is wrong or have a
wrong extension.
}
\examples{
\dontrun{
file<-system.file("extdata","earthquakes_data.txt.zip",package="Capstone")
eq_read_data(file)
}
}
