% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geoms.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
a layer that can be passed ot ggplot containing the time line data
}
\description{
Function that returns the timeline layer to display earthquake data
}
\details{
The function provides a timeline plot of earthquake data.  The size of the plot markers is related
    to the magnitude of the earthquake and the color to the number of reported deaths
    the "x" aesthetic is required and maps to the x-axis of the timeline.  The "y" easthetic is
    optional and can be used to group data by a variable, i.e. country
}
\examples{
\dontrun{
eq_clean \%>\% filter(COUNTRY \%in\% c("JAPAN"), YEAR > 1990) \%>\%
 ggplot(aes(x = DATE,
            y = COUNTRY,
            color = as.numeric(TOTAL_DEATHS),
            size = as.numeric(EQ_PRIMARY)
 )) +
 geom_timeline() +
 theme_timeline() +
 labs(title = "Earthquake Fatalities", color = "# deaths", size = "Richter scale value") +
 scale_color_gradient(low = "black", high = "red")
}

}
