% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Function that add the Eartquakes's Location labels to an timeline earthquake}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, na.rm = TRUE,
  show.legend = NA, stat = "identity", position = "identity",
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{aesthetic mappings}

\item{data}{dataframe that contains the Earthquake's data}

\item{na.rm}{removes the NA values from the data frame}

\item{show.legend}{layer's legend}

\item{stat}{The Layer's statistical transformation}

\item{position}{position adjustment}

\item{inherit.aes}{indicate the default aesthetics}

\item{...}{other arguments}
}
\value{
Plot an Earthquakes timeline which contains the Earthquakes of a country o countries  between two dates
}
\description{
Function that add the Eartquakes's Location labels to an timeline earthquake
}
\examples{
\dontrun{
file<-system.file("extdata","earthquakes_data.txt.zip",package="Capstone")
eq_location_clean(eq_clean_data(eq_read_data(file))) \%>\%
dplyr::filter(DATE >= "1986-02-01" & DATE <="2016-06-01" &
COUNTRY == c("ECUADOR","CHILE", "VENEZUELA"))\%>\%
ggplot() +
geom_timeline(aes(x = DATE, size = EQ_MAG_ML, colour = DEATHS, fill = DEATHS))+
geom_timeline_label(aes(x = DATE, y = COUNTRY, label = LOCATION_NAME,
number = 3, max_aes = EQ_MAG_ML))
}
}
