% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geoms.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeline}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\description{
Implementation of the geom_timeline class
}
\keyword{datasets}
