% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{GeomTimeLineAnnotation}
\alias{GeomTimeLineAnnotation}
\title{Function to add labels on a Earthquake's Location timeline
This Geom will return a plot with the earthquakes of a country or countries between two dates with its respectives names}
\format{An object of class \code{GeomTimeLineAnnotation} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
GeomTimeLineAnnotation
}
\description{
Function to add labels on a Earthquake's Location timeline
This Geom will return a plot with the earthquakes of a country or countries between two dates with its respectives names
}
\keyword{datasets}
