% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geoms.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{GeomTimelineLabel}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimelineLabel
}
\description{
Implementation of the geom_timeline_label class
}
\keyword{datasets}
