% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geoms.R
\name{theme_timeline}
\alias{theme_timeline}
\title{Theme for timeline in ggplot2}
\usage{
theme_timeline()
}
\description{
This is a simple theme that moves the legend to the
   rights so that the deaths labels don't overlap
}
\examples{
\dontrun{
eq_clean \%>\% filter(COUNTRY \%in\% c("GREECE", "ITALY"), YEAR > 2000) \%>\%
 ggplot(aes(x = DATE,
            y = COUNTRY,
            color = as.numeric(TOTAL_DEATHS),
            size = as.numeric(EQ_PRIMARY)
 )) +
 geom_timeline() +
 geom_timeline_label(aes(label = LOCATION_NAME), maxCountries = 5) +
 theme_timeline() +
 labs(title = "Earthquake Fatalities", color = "# deaths", size = "Richter scale value") +
 scale_color_gradient(low = "black", high = "red")
}

}
