% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trimws_to_numeric}
\alias{trimws_to_numeric}
\title{Quite white spaces and then transform to numeric}
\usage{
trimws_to_numeric(x)
}
\arguments{
\item{x}{vector to transform to numeric}
}
\value{
numeric vector
}
\description{
More detailed description
}
\examples{
\dontrun{
x <- c(" 12", "13 ", " 14 ")
trimws_to_numeric(x)
}

}
