% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean Data}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{A dataframe with earthquake data to be cleaned}
}
\value{
The resulting cleaned dataframe
}
\description{
This function cleans the dataframe, creates a DATE column out of the year,
month and day transform the columns LATITUDE and LONGITUD to numeric class
}
\examples{
\dontrun{eq_clean_data(earthquakes_raw)}

}
