% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Function to clean Earthquake dataframe}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{dataframe obtained from the eq_read_data}
}
\value{
dataframe with a new DATE, LATITUDE and LONGITUDE column
}
\description{
Function to clean Earthquake dataframe
}
\examples{
\dontrun{
file<-system.file("extdata","earthquakes_data.txt.zip",package="Capstone")
eq_clean_data(eq_read_data(file))
}

}
