% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean Earthquake data}
\usage{
eq_clean_data(Data)
}
\arguments{
\item{Data}{Earthquake data}
}
\value{
Earthquake data Clean
}
\description{
More detailed description
}
\examples{
\dontrun{
eq_clean_data(Data)
}
}
