% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean NOA Significant Earthquaqe data}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{NOAA Significant Earthquake dataframe}
}
\value{
This function returns a data frame with preprocessed columns: date, longitude
and latitude
}
\description{
This function returns a clean data frame. Preprocessing includes:
date column obtained as a resulted of mergind YEAR, MONTH and DAY columns;
LATITUDE and LONGITUDE columns converted to numeric class
}
\examples{
\dontrun{
df = eq_clean_data(df)
}
}
