% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create Label for Map}
\usage{
eq_create_label(df_create_label = rlang::.data)
}
\arguments{
\item{df_create_label}{A dataframe with earthquake data to create the label}
}
\value{
A column with the structure to be added as a label
}
\description{
This function creates the label to be added to the map plot
}
\examples{
\dontrun{earthquakes_raw \%>\%
eq_clean_data() \%>\%
eq_location_clean() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
mutate(popup_text = eq_create_label(.)) \%>\%
eq_map(annot_col = 'popup_text')}

}
