% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Function that creates a popup text for each earthquake.}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{Clean earthquake data.}
}
\value{
A character vector containing popup text to be used in a leaflet visualization.
}
\description{
This function generates a HTML formatted tag to be used in popups.
}
\examples{
\dontrun{
file<-system.file("extdata","earthquakes_data.txt.zip",package="Capstone")
eq_location_clean(eq_clean_data(eq_read_data(file))) \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 1980) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
 eq_map(name_col = "popup_text")
}

}
