% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creates an HTML label}
\usage{
eq_create_label(DATA)
}
\arguments{
\item{DATA}{Earthquake data}
}
\value{
Earthquake data Clean
}
\description{
More detailed description
}
\examples{
\dontrun{
file_name <- system.file("extdata", "signif.txt", package = "Capstone")
raw_data  <-  file_name \%>\%
       readr::read_tsv() \%>\%
       eq_clean_data() \%>\%
       dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
       dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
       eq_map(annot_col = "popup_text")
}
}
