% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create informative pop-ups for the interactive map created with the eq_map() function}
\usage{
eq_create_label(df_clean)
}
\arguments{
\item{df_clean, }{a dataframe obtained after runing functions eq_clean_data()
and eq_location_clean(). Initial raw data should be downloaded from NOAA website#'}
}
\value{
This function creates HTML label with annotations
}
\description{
This function takes the dataset as an argument and creates an HTML label that
can be used as the annotation text in the leaflet map. This function should put
together a character string for each earthquake that will show the cleaned
location (as cleaned by the eq_location_clean() function created in Module 1),
the magnitude (EQ_PRIMARY), and the total number of deaths (TOTAL_DEATHS), with
boldface labels for each ("Location", "Total deaths", and "Magnitude"). If an
earthquake is missing values for any of these, both the label and the value should
be skipped for that element of the tag.
}
\examples{
\dontrun{
df %>%
  dplyr::filter(COUNTRY == "RUSSIA" & lubridate::year(DATE) >= 2000) %>%
  dplyr::mutate(popup_text = eq_create_label(.)) %>%
  eq_map(annot_col = "popup_text")
  }
}
