% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean Location}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{A dataframe with earthquake data to be cleaned}
}
\value{
The resulting cleaned dataframe
}
\description{
This function cleans the LOCATION column
}
\examples{
\dontrun{eq_clean_data(earthquakes_raw) \%>\% eq_location_clean()}

}
