% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(countryName, location)
}
\arguments{
\item{countryName}{A character string with the country name to be removed}

\item{location}{A character string with the location to be cleaned}
}
\value{
a string with the country and any folowing colons or semi-colons
    removed from the front
}
\description{
Function that removes the country name from the location data
}
\details{
The function first does sanity chaecks on countryName and location
    then checks to see if the location starts with countryName
    if yes, the countryName and trailing colons are removed from the start of location
    str_ucfirst is then used to convert the location to initial upper case
}
\examples{
\dontrun{
eq_location_clean("CANADA", "CANADA: MONTREAL")
}

}
