% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Function for title case the Earthquake's Location Data-Name}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{dataframe with location names written in Uper case}
}
\value{
dataframe with Tittle Case Location names
}
\description{
Function for title case the Earthquake's Location Data-Name
}
\examples{
\dontrun{
file<-system.file("extdata","earthquakes_data.txt.zip",package="Capstone")
eq_location_clean(eq_clean_data(eq_read_data(file)))
}

}
