% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Remove blanks spaces and then transform to numeric}
\usage{
eq_location_clean(LOCATION_NAME)
}
\arguments{
\item{LOCATION_NAME}{vector that include country and location in the same string}
}
\value{
vector that contains only location
}
\description{
More detailed description
}
\examples{
\dontrun{
x <- c("JORDAN:  BAB-A-DARAA,AL-KARAK", "SYRIA:  UGARIT",
       "TURKMENISTAN:  W", "GREECE:  THERA ISLAND (SANTORINI)", "ISRAEL:  ARIHA (JERICHO)" )
eq_location_clean(x)
}

}
