% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean location names for plotting}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{NOAA Significant Earthquake dataframe}
}
\value{
This function returns a data frame with cleaned location names
}
\description{
This function returns a clean data frame.
}
\examples{
\dontrun{
df = eq_location_clean(df)
}
}
