% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_map}
\alias{eq_map}
\title{Map with Earthquakes}
\usage{
eq_map(df_map = rlang::.data, annot_col = NULL)
}
\arguments{
\item{df_map}{A dataframe with earthquake data to be plotted}

\item{annot_col}{The label of the corresponding information for each earthquake}
}
\value{
The resulting map to be plotted
}
\description{
This function plots a World Map indicating each earthquake with a geom point
}
\examples{
\dontrun{earthquakes_raw \%>\%
eq_clean_data() \%>\%
eq_location_clean() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
eq_map()}

}
