% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_map}
\alias{eq_map}
\title{Function that plot the earthquake Data in an Interactive Map.}
\usage{
eq_map(data, name_col)
}
\arguments{
\item{data}{Clean earthquake data.}

\item{name_col}{Name of a column which will be use to show information about the earthquake.}
}
\value{
This function returns an interactive map.
}
\description{
The Earthquakes will be plot using its longitude and latitude information. The user
can select a column in order to obtain information about the earthquake.
}
\note{
If an invalid column name is provided, the function provides a warning
and uses the LOCATION_NAME column as teh annotation column.
}
\examples{
\dontrun{
file<-system.file("extdata","earthquakes_data.txt.zip",package="Capstone")
eq_location_clean(eq_clean_data(eq_read_data(file))) \%>\%
dplyr::filter(COUNTRY == "VENEZUELA" & lubridate::year(DATE) >= 1980) \%>\%
eq_map(name_col = "DATE")
}

}
