% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eq_map}
\alias{eq_map}
\title{Maps the epicenters}
\usage{
eq_map(DATA, annot_col = "DATE", radius = 15)
}
\arguments{
\item{DATA}{Earthquake data}

\item{annot_col}{column with information to show}

\item{radius}{radius circle}
}
\value{
Earthquake data Clean
}
\description{
More detailed description
}
\examples{
\dontrun{
file_name <- system.file("extdata", "signif.txt", package = "Capstone")
raw_data  <-  file_name \%>\%
       readr::read_tsv() \%>\%
       eq_clean_data() \%>\%
       dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
       dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
       eq_map(annot_col = "popup_text")
}
}
