% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{eq_map}
\alias{eq_map}
\title{Mapping the earthquake epicenters
The function maps the epicenters (LATITUDE/LONGITUDE) and annotates each point
with in pop up window containing annotation data stored in a column of the data frame.
The user should be able to choose which column is used for the annotation in the
pop-up with a function argument named annot_col.}
\usage{
eq_map(df_clean, annot_col)
}
\arguments{
\item{df_clean, }{data containing the filtered data frame with earthquakes}

\item{annot_col}{column that should be used for the annotation in the pop-up}
}
\value{
This function returns a leaflet map with earthquake epicentres and annotations
within pop-up window
}
\description{
Mapping the earthquake epicenters
The function maps the epicenters (LATITUDE/LONGITUDE) and annotates each point
with in pop up window containing annotation data stored in a column of the data frame.
The user should be able to choose which column is used for the annotation in the
pop-up with a function argument named annot_col.
}
\examples{
\dontrun{
df %>%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) %>%
  eq_map(annot_col = "DATE")
  }
}
