% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Plot timeline earthquakes}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  na.rm = FALSE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes() or aes_()}

\item{data}{The data to be displayed in this layer. There are three options:
If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().
data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See
fortify() for which variables will be created.
A function will be called with a single argument, the plot data.
The return value must be a data.frame, and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).}

\item{stat}{stat parameter}

\item{position}{position parameter}

\item{show.legend}{show.legend parameter}

\item{na.rm}{na.rm parameter}

\item{inherit.aes}{inherit.aes parameter}

\item{...}{other parameters}
}
\value{
Earthquake data Clean
}
\description{
More detailed description
}
\examples{
\dontrun{
eq_clean_data(Data)
}
}
