% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Plot a timeline of earthquake}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  na.rm = FALSE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created}

\item{data}{The data to be displayed}

\item{stat}{The statistical transformation to use on the data for this layer,
as a string.}

\item{position}{Position adjustment, either as a string, or the result of a
call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than
combining with them.}

\item{...}{Other arguments passed on to the layer. it can be used to set an
aesthetic to a fixed value.}
}
\value{
This function returns a timeline plot
}
\description{
Geom for plotting a time line of earthquakes ranging from xmin to xmaxdates
with a point for each earthquake
}
\examples{
\dontrun{
df %>% filter(COUNTRY %in% c("RUSSIA", "ISRAEL")) %>%
  ggplot(aes(x = date, size = EQ_PRIMARY, y = COUNTRY, col= DEATHS)) +
      geom_timeline() +
      scale_size_continuous(name = 'Richter scale value') +
      scale_color_continuous(name = '# deaths')
}
}
