% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Geom Timeline Proto}
\format{
An object of class \code{GeomTimeline} (inherits from \code{GeomPoint}, \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomTimeline
}
\value{
Visual Structure of the Geom Timeline
}
\description{
Hidden function used to build the Geom Timeline
}
\keyword{internal}
