% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Function to plot an Earthquake's Location timeline
The GeomTimeLine will use a dataframe obtained from the function eq_clean_data.
This Geom will return a plot with the earthquakes of a country or countries between two dates}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimeline
}
\description{
Function to plot an Earthquake's Location timeline
The GeomTimeLine will use a dataframe obtained from the function eq_clean_data.
This Geom will return a plot with the earthquakes of a country or countries between two dates
}
\keyword{datasets}
