% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{New ggproto class for ggplot2: timeline}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimeline
}
\description{
New ggproto class for ggplot2: timeline
}
\keyword{datasets}
