% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-timeline.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{New ggproto class for ggplot2: timelinelabel}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomTimelineLabel
}
\description{
New ggproto class for ggplot2: timelinelabel
}
\keyword{datasets}
